+++ src/main/java/com/atlauncher/gui/dialogs/AddFTBPackDialog.java
@@ -18,6 +18,7 @@
 package com.atlauncher.gui.dialogs;
 
 import java.awt.BorderLayout;
+import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
 import java.awt.GridBagConstraints;
@@ -36,6 +37,7 @@ import javax.swing.JTextField;
 import javax.swing.SwingUtilities;
 
 import com.atlauncher.App;
+import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.constants.Constants;
 import com.atlauncher.data.modpacksch.ModpacksChPackList;
 import com.atlauncher.data.modpacksch.ModpacksChPackManifest;
@@ -43,6 +45,7 @@ import com.atlauncher.gui.card.FTBPackCard;
 import com.atlauncher.gui.layouts.WrapLayout;
 import com.atlauncher.gui.panels.LoadingPanel;
 import com.atlauncher.gui.panels.NoFTBPacksPanel;
+import com.atlauncher.managers.ConfigManager;
 import com.atlauncher.network.Analytics;
 import com.atlauncher.utils.ComboItem;
 
@@ -56,6 +59,7 @@ public final class AddFTBPackDialog extends JDialog {
     private final JPanel topPanel = new JPanel(new BorderLayout());
     private final JTextField searchField = new JTextField(16);
     private final JButton searchButton = new JButton(GetText.tr("Search"));
+    private JLabel platformMessageLabel = null;
     private final JComboBox<ComboItem<String>> sortComboBox = new JComboBox<ComboItem<String>>();
 
     private JScrollPane jscrollPane;
@@ -70,6 +74,11 @@ public final class AddFTBPackDialog extends JDialog {
         this.setPreferredSize(new Dimension(620, 500));
         this.setResizable(false);
         this.setDefaultCloseOperation(JDialog.DISPOSE_ON_CLOSE);
+
+        String platformMessage = ConfigManager.getConfigItem("platforms.modpacksch.message", null);
+        if (platformMessage != null) {
+            platformMessageLabel = new JLabel(new HTMLBuilder().center().text(platformMessage).build());
+        }
 
         sortComboBox.addItem(new ComboItem<>("popular/plays", GetText.tr("Most Popular")));
         sortComboBox.addItem(new ComboItem<>("popular/installs", GetText.tr("Most Installed")));
@@ -110,7 +119,8 @@ public final class AddFTBPackDialog extends JDialog {
         mainPanel.add(this.topPanel, BorderLayout.NORTH);
         mainPanel.add(this.jscrollPane, BorderLayout.CENTER);
 
-        JPanel bottomPanel = new JPanel(new FlowLayout());
+        JPanel bottomPanel = new JPanel(new BorderLayout());
+        JPanel bottomButtonsPanel = new JPanel(new FlowLayout());
 
         prevButton = new JButton("<<");
         prevButton.setEnabled(false);
@@ -120,8 +130,14 @@ public final class AddFTBPackDialog extends JDialog {
         nextButton.setEnabled(false);
         nextButton.addActionListener(e -> goToNextPage());
 
-        bottomPanel.add(prevButton);
-        bottomPanel.add(nextButton);
+        bottomButtonsPanel.add(prevButton);
+        bottomButtonsPanel.add(nextButton);
+
+        if (platformMessageLabel != null) {
+            platformMessageLabel.setForeground(Color.YELLOW);
+            bottomPanel.add(platformMessageLabel, BorderLayout.NORTH);
+        }
+        bottomPanel.add(bottomButtonsPanel, BorderLayout.CENTER);
 
         this.add(mainPanel, BorderLayout.CENTER);
         this.add(bottomPanel, BorderLayout.SOUTH);